Seperate api class objects into seperate files
