+++ cli/src/main/kotlin/com/jmeranda/glazy/cli/commands/PullRequest.kt
@@ -1,0 +1,15 @@
+package com.jmeranda.glazy.cli.commands
+
+import picocli.CommandLine.Command
+
+open class PullCommand
+
+@Command(name = "pull",
+        description = ["Perform operations on pull requests"],
+        mixinStandardHelpOptions = true)
+class PullParent
+
+@Command(name = "list",
+        description = ["List pull requests"],
+        mixinStandardHelpOptions = true)
+class PullList: PullCommand()
