+++ glazy
@@ -9,3 +9,5 @@ if [ ! -f "$jar_file" ]; then
 fi
 
 java -jar "$jar_file" "$@"
+
+exit "$?"
