+++ lib/src/main/kotlin/com/jmeranda/glazy/lib/handler/IssueAllGetHandler.kt
@@ -4,7 +4,7 @@ import khttp.get
 
 import com.fasterxml.jackson.module.kotlin.readValue
 
-import com.jmeranda.glazy.lib.Issue
+import com.jmeranda.glazy.lib.objects.Issue
 import com.jmeranda.glazy.lib.request.IssueGetAllRequest
 
 /**
