+++ lib/src/main/kotlin/com/jmeranda/glazy/lib/objects/Issue.kt
@@ -1,38 +1,4 @@
-package com.jmeranda.glazy.lib
-
-/**
- * Describes a label to an issue.
- */
-data class Label (
-        val id: Int,
-        val nodeId: String,
-        val url: String,
-        val name: String,
-        val color: String,
-        val default: Boolean
-)
-
-/**
- * Describes an issue milestone.
- */
-data class Milestone(
-        val url: String,
-        val htmlUrl: String,
-        val labelsUrl: String,
-        val id: Int,
-        val nodeId: String,
-        val number: Int,
-        val title: String,
-        val description: String,
-        val creator: Owner,
-        val openIssues: Int,
-        val closedIssues: Int,
-        val state: String,
-        val createdAt: String,
-        val updatedAt: String? = null,
-        val dueOn: String? = null,
-        val closedAt: String? = null
-)
+package com.jmeranda.glazy.lib.objects
 
 /**
  * Describes a repository issue.
