+++ lib/src/main/kotlin/com/jmeranda/glazy/lib/objects/Permissions.kt
@@ -1,0 +1,10 @@
+package com.jmeranda.glazy.lib.objects
+
+/**
+ * Describes the permissions on a repository.
+ */
+data class Permissions(
+        val admin: Boolean,
+        val push: Boolean,
+        val pull: Boolean
+)
