+++ lib/src/main/kotlin/com/jmeranda/glazy/lib/service/RepoService.kt
@@ -1,6 +1,6 @@
 package com.jmeranda.glazy.lib.service
 
-import com.jmeranda.glazy.lib.Repo
+import com.jmeranda.glazy.lib.objects.Repo
 import com.jmeranda.glazy.lib.exception.NoSuchRepo
 import com.jmeranda.glazy.lib.handler.*
 import com.jmeranda.glazy.lib.request.*
